TAL-37 - move how skills work modal to shared library; show it in talent search results
