+++ src/libs/shared/lib/components/modals/how-skills-work-modal/HowSkillsWorkModal.module.scss
