+++ src/libs/shared/lib/components/modals/how-skills-work-modal/index.ts
