+++ src/libs/shared/lib/components/modals/index.ts
@@ -2,3 +2,4 @@ export * from './contact-support-modal'
 export * from './order-contract-modal'
 export * from './privacy-policy-modal'
 export * from './terms-modal'
+export * from './how-skills-work-modal'
